+++ app/(service)/categories/[[...categoryName]]/page.tsx
