+++ app/(service)/favorite_goods/page.tsx
