+++ app/(service)/malls/favorite_malls/page.tsx
