+++ app/(service)/products/[product_id]/page.tsx
