+++ app/(service)/user/edit/password/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { PasswordEditForm } from '../../../../components/Form/PasswordEditForm';
+import { PasswordEditForm } from '../../../../../components/Form/PasswordEditForm';
 
 type Props = {};
