+++ app/(service)/user/mysize/page.tsx
@@ -2,8 +2,8 @@ import MobileSection from '@/components/MobileSection';
 import EditButton from '@/components/MyPage/MySize/EditButton';
 import MySizeInfo from '@/components/MyPage/MySize/MySizeInfo';
 import React from 'react';
-import MyPageSection from '../../../components/MyPage/MyPageSection';
-import GuidelinePic from '../../../components/MyPage/MySize/GuidelinePic';
+import MyPageSection from '../../../../components/MyPage/MyPageSection';
+import GuidelinePic from '../../../../components/MyPage/MySize/GuidelinePic';
 
 type Props = {};
