+++ app/layout.tsx
@@ -1,8 +1,6 @@
 import { Inter } from 'next/font/google';
 import type { Metadata } from 'next';
 import './globals.css';
-import Header from '@/components/Header';
-import MobileNavbar from '@/components/Navbar/MobileNavbar';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -17,12 +15,8 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="ko" className={inter.className}>
-      <body className="flex flex-col w-full px-4 lg:px-0 max-w-screen-lg mx-auto mb-[6rem] md:mb-0">
-        <Header />
-        <main className="grow">{children}</main>
-        <MobileNavbar />
-      </body>
+    <html lang="ko">
+      <body className={inter.className}>{children}</body>
     </html>
   );
 }
